Changed camera clipping behaviour
Previously I literally had a rigidbody and collider attached to the
camera to physically simulate collisions between the terrain and the
camera, but that caused the camera to jump out of the terrain in other
directions so that when it touched the terrain it would kind of jiggle
around instead of just pointing directly at the player at all times. I
changed it today to just take the normal of the raycast hit and adjust
the camera's final position according to that. It never even crossed my
mind before, but today I was thinking about the problems I was having
with the rigidbody + collider method and how I could get the camera to
always move away from the surface that it was touching, and then I
thought "wait, that's what a normal is, I wonder if I can get the normal
of a raycast", and it turned out I could get the normal of a raycast, so
I did, and it worked pretty well, but I'm sure it could be improved by
taking into account the angle of the camera and the distance in relation
to the normal, or some other stuff that I didn't think too hard about.

There's still issues with the camera clipping into surfaces when the
distance of the ray is greater than the max camera distance, but this is
probably good enough for now.